Refactoring away from FastEndpoints test classes
TestBase class was causing problems with the database being disposed between test classes.  Tried refactoring to use collections, but it failed.  Also tried refactoring to place the db fixture in FastEndpoints state, but it had problems with initialization order that did not work with the dependency between the web app and database.  Ultimately it was easiest to move away from the FE helpers and explicitly work with XUnit to have full control over the lifetime and initialization.